requirements.txt
boto3==1.24.15
google-cloud-storage==2.10.0
requests==2.26.0
